requirements.txt
casvisor-python-sdk~=1.0.0

